Cleaned up code style issue and comment style issues. Renamed the annotation builder to something more appropriate. Moved the annotation builder and scala test util class into the impl package.
