+++ src/java/org/pantsbuild/tools/junit/impl/CustomAnnotationBuilder.java
@@ -1,7 +1,7 @@
 // Copyright 2015 Pants project contributors (see CONTRIBUTORS.md).
 // Licensed under the Apache License, Version 2.0 (see LICENSE).
 
-package org.pantsbuild.tools.junit.withretry;
+package org.pantsbuild.tools.junit.impl;
 
 import java.io.PrintStream;
 
@@ -10,17 +10,19 @@ import org.junit.internal.builders.AnnotatedBuilder;
 import org.junit.internal.builders.JUnit4Builder;
 import org.junit.runner.Runner;
 import org.junit.runners.model.RunnerBuilder;
+import org.pantsbuild.tools.junit.withretry.JUnit4BuilderWithRetry;
 
 /**
- * Needed to support retrying flaky tests. Using method overriding, gives us access to code
- * in JUnit4 that cannot be customized in a simpler way.
+ * Needed to support retrying flaky tests as well as add support for running scala tests.
+ * Using method overriding, gives us access to code in JUnit4 that cannot be customized
+ * in a simpler way.
  */
-public class AllDefaultPossibilitiesBuilderWithRetry extends AllDefaultPossibilitiesBuilder {
+public class CustomAnnotationBuilder extends AllDefaultPossibilitiesBuilder {
 
   private final int numRetries;
   private final PrintStream err;
 
-  public AllDefaultPossibilitiesBuilderWithRetry(int numRetries, PrintStream err) {
+  public CustomAnnotationBuilder(int numRetries, PrintStream err) {
     super(true);
     this.numRetries = numRetries;
     this.err = err;
