+++ tests/java/org/pantsbuild/tools/junit/lib/BUILD
@@ -14,3 +14,4 @@ java_library(
   ],
   sources=globs('*.java')
 )
+''
