+++ tests/scala/org/pantsbuild/tools/junit/lib/MockScalaTest.scala
@@ -2,9 +2,6 @@ package org.pantsbuild.tools.junit.lib
 
 import org.scalatest.FreeSpec
 
-/**
-  * Created by dbrewster on 11/2/16.
-  */
 class MockScalaTest extends FreeSpec {
   "test" - {
     "should pass" in {
