+++ tests/scala/org/pantsbuild/tools/junit/lib/NotAScalaTest.scala
@@ -1,8 +1,3 @@
 package org.pantsbuild.tools.junit.lib
 
-/**
-  * Created by dbrewster on 11/2/16.
-  */
-class NotAScalaTest {
-
-}
+class NotAScalaTest {}
